conditionally show WobblyRow
