+++ catalogue/webapp/pages/work.js
@@ -5,7 +5,7 @@ import {
   type CatalogueApiError,
   type CatalogueApiRedirect,
 } from '@weco/common/model/catalogue';
-import { useEffect, useState, type Node } from 'react';
+import { useEffect, useState } from 'react';
 import fetch from 'isomorphic-unfetch';
 import { spacing, grid, classNames } from '@weco/common/utils/classnames';
 import {
@@ -33,25 +33,7 @@ import TogglesContext from '@weco/common/views/components/TogglesContext/Toggles
 import MessageBar from '@weco/common/views/components/MessageBar/MessageBar';
 import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
 import BetaMessage from '@weco/common/views/components/BetaMessage/BetaMessage';
-
-type WobblyProps = {|
-  children: Node,
-|};
-
-const WobblyRow = ({ children }: WobblyProps) => (
-  <div
-    className={classNames({
-      'row bg-cream row--has-wobbly-background': true,
-    })}
-  >
-    <div className="container">
-      <div className="grid">
-        <div className={grid({ s: 12, m: 12, l: 12, xl: 12 })}>{children}</div>
-      </div>
-    </div>
-    <div className="row__wobbly-background" />
-  </div>
-);
+import WobblyRow from '@weco/common/views/components/WobblyRow/WobblyRow';
 
 type Props = {|
   work: Work | CatalogueApiError,
@@ -222,39 +204,35 @@ export const WorkPage = ({ work }: Props) => {
         {({ showMultiVolumePreviews }) => (
           <>
             {!showMultiVolumePreviews && iiifPresentationManifests ? (
-              <WobblyRow>
-                <div
-                  className={classNames({
-                    [spacing({ s: 4 }, { margin: ['bottom'] })]: true,
-                  })}
-                >
-                  <BetaMessage message="We are working to make this item available online in July 2019." />
-                </div>
-              </WobblyRow>
+              <div
+                className={classNames({
+                  [spacing({ s: 4 }, { margin: ['bottom'] })]: true,
+                })}
+              >
+                <BetaMessage message="We are working to make this item available online in July 2019." />
+              </div>
             ) : (
               iiifPresentationManifests &&
               iiifPresentationManifests.map((manifest, i) => (
                 <ManifestContext.Provider value={manifest} key={i}>
                   <SpacingComponent>
-                    <WobblyRow>
-                      {sierraIdFromPresentationManifestUrl &&
-                        !iiifImageLocationUrl && (
-                          <IIIFPresentationPreview
-                            iiifPresentationLocation={iiifPresentationLocation}
-                            itemUrl={itemUrl({
-                              workId: work.id,
-                              sierraId:
-                                manifest['@id'].match(
-                                  /^https:\/\/wellcomelibrary\.org\/iiif\/(.*)\/manifest$/
-                                )[1] || sierraIdFromPresentationManifestUrl,
-                              langCode: work.language && work.language.id,
-                              page: 1,
-                              canvas: 1,
-                              isOverview: true,
-                            })}
-                          />
-                        )}
-                    </WobblyRow>
+                    {sierraIdFromPresentationManifestUrl &&
+                      !iiifImageLocationUrl && (
+                        <IIIFPresentationPreview
+                          iiifPresentationLocation={iiifPresentationLocation}
+                          itemUrl={itemUrl({
+                            workId: work.id,
+                            sierraId:
+                              manifest['@id'].match(
+                                /^https:\/\/wellcomelibrary\.org\/iiif\/(.*)\/manifest$/
+                              )[1] || sierraIdFromPresentationManifestUrl,
+                            langCode: work.language && work.language.id,
+                            page: 1,
+                            canvas: 1,
+                            isOverview: true,
+                          })}
+                        />
+                      )}
                   </SpacingComponent>
                 </ManifestContext.Provider>
               ))
@@ -264,23 +242,21 @@ export const WorkPage = ({ work }: Props) => {
       </TogglesContext.Consumer>
 
       <ManifestContext.Provider value={iiifPresentationManifest}>
-        {!iiifPresentationManifests && (
-          <WobblyRow>
-            {sierraIdFromPresentationManifestUrl && !iiifImageLocationUrl && (
-              <IIIFPresentationPreview
-                iiifPresentationLocation={iiifPresentationLocation}
-                itemUrl={itemUrl({
-                  workId: work.id,
-                  sierraId: sierraIdFromPresentationManifestUrl,
-                  langCode: work.language && work.language.id,
-                  page: 1,
-                  canvas: 1,
-                  isOverview: true,
-                })}
-              />
-            )}
-          </WobblyRow>
-        )}
+        {!iiifPresentationManifests &&
+          sierraIdFromPresentationManifestUrl &&
+          !iiifImageLocationUrl && (
+            <IIIFPresentationPreview
+              iiifPresentationLocation={iiifPresentationLocation}
+              itemUrl={itemUrl({
+                workId: work.id,
+                sierraId: sierraIdFromPresentationManifestUrl,
+                langCode: work.language && work.language.id,
+                page: 1,
+                canvas: 1,
+                isOverview: true,
+              })}
+            />
+          )}
       </ManifestContext.Provider>
 
       {iiifImageLocationUrl && (
